# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Starter Kit API",
        "description": "API documentation for the Laravel Filament Starter Kit.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Current application environment"
        }
    ],
    "paths": {
...
        "/api/v1/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Daftar postingan publik",
                "operationId": "7b253f6aade6833019f938c1f2ef08ed",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Kata kunci pencarian",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_slug",
                        "in": "query",
                        "description": "Filter berdasarkan slug kategori",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filter berdasarkan slug atau ID tag",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Filter berdasarkan username/email penulis",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar post berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PostResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Buat post baru",
                "operationId": "b996500c6caf55100cd024d3e3b49dd5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostResource"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    },
                    "403": {
                        "description": "Tidak memiliki izin"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/posts/{post}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Detail post",
                "operationId": "a8c7efde35a86efcbbf0bfbfc04b10be",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Slug post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail post berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Perbarui post",
                "operationId": "6c7a852cde873d1dccbac9d2ac4cb0f5",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Slug post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostResource"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    },
                    "403": {
                        "description": "Tidak memiliki izin"
                    },
                    "404": {
                        "description": "Post tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Hapus post",
                "operationId": "c8a5edc6ae443a76b3c3ea3abeeb6d54",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Slug post",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post dihapus"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    },
                    "403": {
                        "description": "Tidak memiliki izin"
                    },
                    "404": {
                        "description": "Post tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CategoryResource": {
                "title": "Category",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "name": {
                        "type": "string",
                        "example": "News"
                    },
                    "slug": {
                        "type": "string",
                        "example": "news"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "image": {
                        "type": "string",
                        "nullable": true
                    },
                    "icon": {
                        "type": "string",
                        "nullable": true
                    },
                    "color": {
                        "type": "string",
                        "example": "#F59E0B",
                        "nullable": true
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_featured": {
                        "type": "boolean"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "posts_count": {
                        "type": "integer",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "PostResource": {
                "title": "Post",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "title": {
                        "type": "string",
                        "example": "How to use the starter kit"
                    },
                    "slug": {
                        "type": "string",
                        "example": "how-to-use-the-starter-kit"
                    },
                    "excerpt": {
                        "type": "string",
                        "nullable": true
                    },
                    "content": {
                        "type": "string"
                    },
                    "featured_image": {
                        "type": "string",
                        "nullable": true
                    },
                    "gallery": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "type": {
                        "type": "string",
                        "example": "article"
                    },
                    "status": {
                        "type": "string",
                        "example": "published"
                    },
                    "is_featured": {
                        "type": "boolean"
                    },
                    "is_sticky": {
                        "type": "boolean"
                    },
                    "view_count": {
                        "type": "integer",
                        "example": 150
                    },
                    "reading_time": {
                        "type": "integer",
                        "example": 4
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "scheduled_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "seo": {
                        "properties": {
                            "title": {
                                "type": "string",
                                "nullable": true
                            },
                            "description": {
                                "type": "string",
                                "nullable": true
                            },
                            "keywords": {
                                "type": "string",
                                "nullable": true
                            },
                            "og_image": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategoryResource"
                    },
                    "author": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TagResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TagResource": {
                "title": "Tag",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 4
                    },
                    "name": {
                        "type": "string",
                        "example": "Laravel"
                    },
                    "slug": {
                        "type": "string",
                        "example": "laravel"
                    },
                    "type": {
                        "type": "string",
                        "example": "post"
                    },
                    "color": {
                        "type": "string",
                        "example": "#2563EB",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin User"
                    },
                    "username": {
                        "type": "string",
                        "example": "admin"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@example.com"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "https://...",
                        "nullable": true
                    },
                    "bio": {
                        "type": "string",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "preferences": {
                        "type": "object",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "last_login_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}